custom_components/heizoel24mex/__init__.py
from __future__ import annotations

from homeassistant.config_entries import ConfigEntry
from homeassistant.const import Platform
from homeassistant.core import HomeAssistant

from .const import CONF_PASSWORD, CONF_USERNAME, DOMAIN
from .sensor import Heizoel24MexCoordinator, Heizoel24MexData

# TODO List the platforms that you want to support.
# For your initial PR, limit it to 1 platform.
PLATFORMS: list[Platform] = [Platform.SENSOR]


async def async_setup_entry(hass: HomeAssistant, config: ConfigEntry) -> bool:
    """Set up the element."""
    hass.data.setdefault(DOMAIN, {})

    data = Heizoel24MexData(
        config.data.get(CONF_USERNAME), config.data.get(CONF_PASSWORD)
    )
    coordinator: Heizoel24MexCoordinator = Heizoel24MexCoordinator(hass, data)
    await coordinator.async_config_entry_first_refresh()
    hass.data.setdefault(DOMAIN, {})[config.entry_id] = coordinator

    # Add entities
    await hass.config_entries.async_forward_entry_setups(config, PLATFORMS)

    return True


async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
    """Unload a config entry."""
    if unload_ok := await hass.config_entries.async_unload_platforms(entry, PLATFORMS):
        hass.data[DOMAIN].pop(entry.entry_id)

    return unload_ok
