requirements.test.txt
pytest
pytest-cov
pytest-homeassistant-custom-component
